Oh, say, can you see, by the dawn's early light, What so proudly we hailed at the twilight's last gleaming? Whose broad stripes and bright stars, thro' the perilous fight' O'er the ramparts we watched, were so gallantly streaming. And the rockets red glare, the bombs bursting in air, Gave proof through the night that our flag was still there. Oh, say, does that star-spangled banner yet wave O'er the land of the free and the home of the brave?
On the shore dimly seen, thro' the mists of the deep,
Where the foe's haughty host in dread silence reposes,
What is that which the breeze, o'er the towering steep,
As it fitfully blows, half conceals, half discloses?
Now it catches the gleam of the morning's first beam,
In full glory reflected, now shines on the stream;
'Tis the star-spangled banner: oh, long may it wave
O'er the land of the free and the home of the brave.

And where is that band who so vauntingly swore
That the havoc of war and the battle's confusion
A home and a country should leave us no more?
Their blood has wash'd out their foul footstep's pollution.
No refuge could save the hireling and slave
From the terror of flight or the gloom of the grave,
And the star-spangled banner in triumph doth wave
O'er the land of the free and the home of the brave.

Oh, thus be it ever when free men shall stand,
Between their loved homes and the war's desolation;
Blest with vict'ry and peace, may the heav'n-rescued land
Praise the Power that has made and preserved us as a nation.
Then conquer we must, when our cause is just,
And this be our motto: "In God is our trust";
And the star-spangled banner in triumph shall wave
O'er the land of the free and the home of the brave.